requirements.txt
fastapi
uvicorn
openai
python-dotenv
streamlit
pytesseract
pillow
PyMuPDF
